% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{avg_gwp_rate}
\alias{avg_gwp_rate}
\title{Compute the average per-capita gross world product growth rate over the
 specified years.}
\usage{
avg_gwp_rate(pop, gwp, yrs, start = 2010, end = 2100)
}
\arguments{
\item{pop}{Numeric vector of global population.}

\item{gwp}{Numeric vector of gross world products.}

\item{yrs}{Numeric vector of years corresponding to the values in pop and
gwp.}

\item{start}{Starting year for the average per-capita GWP growth rate
calculation.}

\item{end}{Ending year for the average per-capita GWP growth rate
calculation.}
}
\value{
Average per-capita GWP growth rate value (on a decimal scale, not
 a percentage scale.)
}
\description{
\code{avg_gwp_rate} computes the average per-capita gross world product
 (GWP) growth rate for a model simulation over the specified years.
}
