% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{check_fossil_constraint}
\alias{check_fossil_constraint}
\title{Check if model output satisfies the fossil fuel resource constraint.}
\usage{
check_fossil_constraint(model_out, start = 1700, end = 2500, thresh = NA)
}
\arguments{
\item{model_out}{Data frame produced by the model.}

\item{start}{First year for the fossil constraint.}

\item{end}{Last year for the fossil constraint.}

\item{thresh}{Fossil fuel resource constraint value (in Gt C).}
}
\value{
Boolean value corresponding to if the model output satisfies the
 constraint.
}
\description{
\code{check_fossil_constraint} checks if the emissions output is within
 the fossil fuel resource limit.
}
\details{
This model assumes that there is a finite limit of Gt C that are
 associated with fossil fuel emissions. \code{check_fossil_constraint}
 checks if that constraint is satisfied by the model output.
}
