% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{check_penetration_constraints}
\alias{check_penetration_constraints}
\title{Check the tech penetration constraints.}
\usage{
check_penetration_constraints(model_out, years = NA, windows = NA)
}
\arguments{
\item{model_out}{Data frame produced by the model.}

\item{years}{Numeric vector of years for which the constraint is defined. By default this is \code{NA}, as the constraint is turned off. In principle the function should not be called if \code{years=NA}.}

\item{windows}{List specifying the windows for each technology at each year. Each element of the list should correspond to a year set in \code{years}. The windows should be specified in a 2x3 matrix where the rows are the lower and upper bounds and the columns correspond to the high-emitting technology, the low-emitting technology, and the zero-emissions technology. If a particular technology is not used in the constraint, set the corresponding row values as \code{NA}. The pre-industrial technology is ignored. By default this has a value of \code{NA}, as the constraint is turned off. An error will be returned if the format of this list doesn't match the specification and the number of years.}
}
\value{
Boolean value corresponding to whether the simulation passes the constraint.
}
\description{
\code{check_penetration_constraints} check whether the simulation results pass the technological penetration constraints.
}
\details{
The default priors for the technological substitution process are quite broad, so this check helps to constrain the range of outcomes to reflect more realistic dynamics.
}
