% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_simulate.R
\name{cond_sim_model}
\alias{cond_sim_model}
\title{Conditional simulation of out-of-sample data.}
\usage{
cond_sim_model(pars, parnames, dat, projyrs = NULL, hoyrs = NULL)
}
\arguments{
\item{pars}{Numeric vector of parameter values. These parameters must
include all of the model parameters listed in Table S1 and the
statistical parameters listed in Table S2 of Srikrishnan &
Keller (2019), with Greek letters spelled out: \itemize{
\item psi1, the population growth rate;
\item psi2, the population half-saturation constant;
\item psi3, the population carrying capacity;
\item P0, the initial population in year \code{start};
\item lambda, the elasticity of production with respect to labor (must
be less than 1);
\item s, the savings rate;
\item delta, the capital depreciation rate (must be less than s);
\item alpha, the rate of technological progress for total factor
productivity;
\item As, the saturation level of total factor productivity;
\item pi, the labor participation rate (must be less than 1);
\item A0, the initial total factor productivity in year \code{start};
\item rho2, the carbon emissions intensity of technology 2;
\item rho3, the carbon emissions intensity of technology 3;
\item tau2, the half-saturation year of technology 2;
\item tau3, the half-saturation year of technology 3;
\item tau4, the half-saturation year of technology 4;
\item kappa, the rate of technological penetration;
\item aij, for $i,j=1, 2, 3$, the elements of the VAR cefficient matrix
(only when using a VAR model for the likelihood structure);
\item sigma_pop, the variance of the VAR or normal innovations for
population;
\item sigma_prod, the variance of the VAR or normal innovations for
economic output;
\item sigma_emis, the variance of the VAR or normal innovations for
emissions;
\item eps_pop, the variance of the observation errors for population;
\item eps_prod, the variance of the observation errors for economic
output;
\item eps_emis, the variance of the observation errors for emissions;
}}

\item{parnames}{Character vector of parameter names. These names should
align with the values in \code{pars}, but they don't need to be in any
particular order otherwise.}

\item{dat}{List of data frames of data. List should have names 'pop'
(population), 'prod' (production), and 'emissions,' and each data frame
should have two columns, 'year' and 'value'.}

\item{projyrs}{Vector of years for projections. Should start in 2015 and
be contiguous.}

\item{hoyrs}{Vector of years to hold out. These can be any years contained
in the data.}
}
\value{
Data frame of model simulation, with columns 'year, 'P'
 (population), 'Q' (economic output), and 'C' (CO2 emissions). Only the
 held-out and projected simulation years are returned.
}
\description{
\code{cond_sim_model} simulates out-of-sample data (held-out or
 projections) conditional on the observations.
}
\details{
The joint likelihood function for a VAR(1) model is a multivariate
 Gaussian. \code{cond_sim_model} partitions the data into observations,
 which are treated as fixed, held-out data, and future projections. The
 model discrepancies of the simulation follows the desired joint
 likelihood. The variance of the held-out data takes into account
 observation error variances, while that is not included for projections.

At least one of projyrs (vector of years for projections) and hoyrs
 (vector of years for held-out data) must be passed. It is ok for both to
 be passed.
}
