% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_mat}
\alias{cov_mat}
\title{Construct the joint covariance matrix.}
\usage{
cov_mat(A, B, D, H)
}
\arguments{
\item{A}{Numeric VAR(1) coefficient matrix.}

\item{B}{Numeric covariance matrix for the observations, without
observation errors.}

\item{D}{Numeric covariance matrix for the observation errors.}

\item{H}{Integer matrix giving the lags corresponding to each block.}
}
\value{
Joint covariance matrix, including observation error variances.
}
\description{
The joint covariance matrix for the VAR(1) model is constructed by
  joining blocks based on lags.
}
