% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{create_prior_list}
\alias{create_prior_list}
\title{Create list of prior density information based on input data frame.}
\usage{
create_prior_list(prior_df)
}
\arguments{
\item{prior_df}{Data frame of information about the prior density
functions. Each row (corresponding to a single parameter) should have the
following entries columns:\itemize{
\item name, the parameter name;
\item type, the type of prior density function. Options are "normal,"
"uniform," and "log-normal";
\item lower, numeric value defining the lower value. If the type is
"normal," this should be the 0.05 quantile (or the lower bound of the 90%
credible interval). If the type is "uniform," this should be the lower
bound of the distribution. If the log-normal, this value is ignored;
\item upper, numeric value defining the upper value. If the type is
"normal," this should be the 0.95 quantile (or the upper bound of the 90%
credible interval). If the type is "uniform," this should be the upper
bound of the distribution. If the log-normal, this value is ignored.
}}
}
\value{
List with each entry corresponding to an individual parameter.
 Each of these lists has the following entries: \itemize{
 \item type, the parameter type;
 \item dens.fun, string giving the name of the function which evaluates
 the density;
 \item quant.fun, string giving the name of the function which computes
 quantiles;
 \item rand.fun, string giving the name of the function which generates
 random samples;
 \item other parameters required for the various function calls, which
 depend on the type of distribution.
 }
}
\description{
\code{create_prior_list} creates a list for use in \code{\link{log_pri}}.
 The parameters for each prior density function are estimated from the
 confidence interval bounds in the data frame. Available (by default)
 density functions are normal, uniform, and log-normal. Log-normal priors
 are defined by meanlog = -1 and sdlog=1 no matter what.
}
