% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cum_co2}
\alias{cum_co2}
\title{Compute the cumulative CO2 emissions over the specified years.}
\usage{
cum_co2(emis, yrs, start = 2000, end = 2100)
}
\arguments{
\item{emis}{Numeric vector of carbon emissions.}

\item{yrs}{Numeric vector of years corresponding to the values in pop and
gwp.}

\item{start}{Starting year for the average per-capita GWP growth rate
calculation.}

\item{end}{Ending year for the average per-capita GWP growth rate
calculation.}
}
\value{
Cumulative CO2 emissions (in Gt C) over the specified period.
}
\description{
\code{cum_co2} computes the cumulative CO2 emissions for a model
 simulation over the specified years.
}
