% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{find_map}
\alias{find_map}
\title{Find the maximum a posteriori (MAP) estimate for the IAM.}
\usage{
find_map(
  post,
  parnames,
  residtype,
  prior_df,
  data_yrs,
  n_iter = 5000,
  NP_scale = 25,
  parallel = TRUE,
  trace = FALSE,
  ...
)
}
\arguments{
\item{post}{Function call (not string) for the log-posterior density
function.}

\item{parnames}{Character vector of parameter names (see the documentation
for the desired likelihood function for which parameters are required).}

\item{residtype}{String for the type of residual structure (by default
"iid" or "var"). The log-likelihood function associated with the
structure should have the name log_lik_residtype.}

\item{prior_df}{Data frame of prior information (such as that produced by
\code{\link{create_prior_list}}).}

\item{data_yrs}{Numeric vector of the years for which the data will be
assimilated.}

\item{n_iter}{Number of iterations for \code{\link{DEoptim}}.}

\item{NP_scale}{Numeric factor for the number of particles (relative to
the number of parameters).}

\item{parallel}{Boolean: Should \code{\link{DEoptim}} be run in parallel?}

\item{trace}{Boolean: Should \code{\link{DEoptim}} print its trace?}

\item{...}{Other parameters passed to the log-posterior function.}
}
\value{
Output of \code{\link{DEoptim}} call. The best parameter vector is
 named using parnames.
}
\description{
\code{find_map} finds a MAP estimate using a differential evolution
 function (in this case, \code{DEoptim}).
}
