% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{iid.sim}
\alias{iid.sim}
\title{Simulate white noise.}
\usage{
iid.sim(N, sigma)
}
\arguments{
\item{N}{Length of white noise sequence.}

\item{sigma}{Standard deviation for the white noise.}
}
\value{
Vector of generated white noise samples.
}
\description{
\code{iid.sim} is a wrapper for simulating white noise. The samples are
 drawn from a univariate normal distribution with mean zero.
}
