% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{log_exp_co2}
\alias{log_exp_co2}
\title{Log-density for the expert assessment of CO2 emissions in 2100.}
\usage{
log_exp_co2(model_out)
}
\arguments{
\item{model_out}{Data frame of model output (from \code{\link{run_model}
}).}
}
\value{
Log-density value of the model average GWP per-capita growth rate.
}
\description{
\code{log_exp_co2} returns the log-density for the expert assessment of
   CO2 emissions in 2100 from Ho et al (2019), which is used in
   probabilistic inversion. This
}
\details{
This function uses a normal distribution fit to the median 25% and 75%
 percentiles of experts who did not include negative emissions levels in
 their central 90% interval. It evaluates the density value of the model
 output based on this reference distribution.
}
