% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{log_exp_pop}
\alias{log_exp_pop}
\title{Log-density for the expert assessment of population in 2100.}
\usage{
log_exp_pop(model_out)
}
\arguments{
\item{model_out}{Data frame of model output (from \code{\link{run_model}
}).}
}
\value{
Log-density value of the model population in 2100.
}
\description{
\code{log_exp_pop} returns the log-density for the probabilistic projections of
   population in 2100 from the United Nations, Department of Economic and Social Affairs,
   Population Divison (2019), which is used in
   probabilistic inversion.
}
\details{
This function uses a normal distribution fit to the 95% prediction interval.
It evaluates the density value of the model
 output based on this reference distribution.
}
