% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{log_lik_var}
\alias{log_lik_var}
\title{Compute the log-likelihood under the assumption of residuals jointly
 distributed according to a VAR(1) model.}
\usage{
log_lik_var(pars, parnames, model_out, dat, hoyrs = NULL)
}
\arguments{
\item{pars}{Numeric vector of parameter values. These parameters must
include all of the model parameters listed in Table S1 and the
statistical parameters listed in Table S2 of Srikrishnan &
Keller (2019), with Greek letters spelled out: \itemize{
\item psi1, the population growth rate;
\item psi2, the population half-saturation constant;
\item psi3, the population carrying capacity;
\item P0, the initial population in year \code{start};
\item lambda, the elasticity of production with respect to labor (must
be less than 1);
\item s, the savings rate;
\item delta, the capital depreciation rate (must be less than s);
\item alpha, the rate of technological progress for total factor
productivity;
\item As, the saturation level of total factor productivity;
\item pi, the labor participation rate (must be less than 1);
\item A0, the initial total factor productivity in year \code{start};
\item rho2, the carbon emissions intensity of technology 2;
\item rho3, the carbon emissions intensity of technology 3;
\item tau2, the half-saturation year of technology 2;
\item tau3, the half-saturation year of technology 3;
\item tau4, the half-saturation year of technology 4;
\item kappa, the rate of technological penetration;
\item aij, for $i,j=1, 2, 3$, the elements of the VAR cefficient matrix;
\item sigma_pop, the variance of the VAR innovations for population;
\item sigma_prod, the variance of the VAR innovations for economic output
;
\item sigma_emis, the variance of the VAR innovations for emissions;
\item eps_pop, the variance of the observation errors for population;
\item eps_prod, the variance of the observation errors for economic
output;
\item eps_emis, the variance of the observation errors for emissions.
}}

\item{parnames}{Character vector of parameter names. These names should
align with the values in \code{pars}, but they don't need to be in any
particular order otherwise.}

\item{model_out}{Data frame produced by model run, with columns 'year',
P' (population), 'Q' (production), and 'C' (emissions).}

\item{dat}{List of data frames of data. List should have names 'pop'
(population), 'prod' (production), and 'emissions,' and each data frame
should have two columns, 'year' and 'value'.}

\item{hoyrs}{Integer vector of years that should be held out from the
likelihood function computation.}
}
\value{
Numeric value for the log-likelihood of the parameters given the
 data and the fossil fuel constraint value.
}
\description{
\code{log_lik_var} assumes that the model residuals are jointly
 distributed according to a vector autoregressive model of order 1 (VAR(1
 )). It also checks if the model output satisfies a fossil fuel constraint
 , and returns -infinity if it does not.
}
