% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{log_pri}
\alias{log_pri}
\title{Log-prior density for the provided parameter values}
\usage{
log_pri(pars, parnames, priors)
}
\arguments{
\item{pars}{Numeric vector of parameter values. These parameters must
include all of the model parameters listed in Table S1 and the
statistical parameters listed in Table S2 of Srikrishnan &
Keller (2019), with Greek letters spelled out: \itemize{
\item psi1, the population growth rate;
\item psi2, the population half-saturation constant;
\item psi3, the population carrying capacity;
\item P0, the initial population in year \code{start};
\item lambda, the elasticity of production with respect to labor (must
be less than 1);
\item s, the savings rate;
\item delta, the capital depreciation rate (must be less than s);
\item alpha, the rate of technological progress for total factor
productivity;
\item As, the saturation level of total factor productivity;
\item pi, the labor participation rate (must be less than 1);
\item A0, the initial total factor productivity in year \code{start};
\item rho2, the carbon emissions intensity of technology 2;
\item rho3, the carbon emissions intensity of technology 3;
\item tau2, the half-saturation year of technology 2;
\item tau3, the half-saturation year of technology 3;
\item tau4, the half-saturation year of technology 4;
\item kappa, the rate of technological penetration;
\item aij, for $i,j=1, 2, 3$, the elements of the VAR cefficient matrix
(only when using a VAR model for the likelihood structure);
\item sigma_pop, the variance of the VAR or normal innovations for
population;
\item sigma_prod, the variance of the VAR or normal innovations for
economic output;
\item sigma_emis, the variance of the VAR or normal innovations for
emissions;
\item eps_pop, the variance of the observation errors for population
(only when using a VAR model for the likelihood structure);
\item eps_prod, the variance of the observation errors for economic
output (only when using a VAR model for the likelihood structure);
\item eps_emis, the variance of the observation errors for emissions.
(only when using a VAR model for the likelihood structure);
}}

\item{parnames}{Character vector of parameter names. These names should
align with the values in \code{pars}, but they don't need to be in any
particular order otherwise.}

\item{priors}{Named list of priors, with names containing all of the
elements in \code{parnames}. \code{priors} is best created using
'\code{\link{create_prior_list}}, but each element must be a list
containing: \itemize{
\item dens.fun, (string) function name for evaluating the density
\item ..., other named parameters required for evaluation of
\code{dens.fun}
}}
}
\value{
Joint log-prior density value parameters based on the given
 priors.
}
\description{
\code{log_pri} returns the log-prior density for the provided parameter
 values, based on the provided list of prior distributions.
}
\details{
The provided prior distributions (which should be in the list form
 generated by \code{\link{create_prior_list}}) are treated as independent
 prior distributions, so that the joint log-prior density is the sum of
 the individual log-prior densities.
}
