% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{model_run}
\alias{model_run}
\title{Run the IAM.}
\usage{
model_run(yr, P0, psi, alpha, A0, As, s, lambda, delta, pi, kappa, gamma,
  rho, init)
}
\arguments{
\item{yr}{Numeric vector of years for the model run.}

\item{P0}{Double, the initial population.}

\item{psi}{Numeric vector with three components for the population
dynamics variables psi1 (growth rate), psi2 (half-saturation constant),
and psi3 (carrying capacity).}

\item{alpha}{Double, the rate of technological progress for total
factor productivity.}

\item{A0}{Double, the initial total factor productivity.}

\item{As}{Double, the total factor productivity saturation level.}

\item{s}{Double, the savings rate (must be less than 1).}

\item{lambda}{Double, the elasticity of production with respect to labor.}

\item{delta}{Double,the capital depreciation rate (must be less than 1).}

\item{pi}{Double, the labor force participation rate (must be less than 1
).}

\item{kappa}{Double, the rate of technological penetration (must be less
than 1).}

\item{gamma}{Numeric matrix, each column is the penetration for the
corresponding technology, and each row corresponds to a model year.}

\item{rho}{Numeric vector for the carbon intensities of each technology.}

\item{init}{Optional initial values for Q, and C.}
}
\value{
Data frame of model output.
}
\description{
Run the IAM.
}
