% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{neg_log_post}
\alias{neg_log_post}
\title{Negative log-posterior density for the provided parameter values given the
 data.}
\usage{
neg_log_post(
  pars,
  parnames,
  priors,
  dat,
  lik_fun,
  exp_gwp = FALSE,
  exp_co2 = FALSE,
  exp_pop = FALSE,
  ff_thresh = NA,
  ff_const_yrs = NA,
  ff_pen_windows = NA,
  ff_pen_yrs = NA,
  hoyrs = NULL
)
}
\arguments{
\item{pars}{Numeric vector of parameter values. These parameters must
include all of the model parameters listed in Table S1 and the
statistical parameters listed in Table S2 of Srikrishnan &
Keller (2019), with Greek letters spelled out: \itemize{
\item psi1, the population growth rate;
\item psi2, the population half-saturation constant;
\item psi3, the population carrying capacity;
\item P0, the initial population in year \code{start};
\item lambda, the elasticity of production with respect to labor (must
be less than 1);
\item s, the savings rate;
\item delta, the capital depreciation rate (must be less than s);
\item alpha, the rate of technological progress for total factor
productivity;
\item As, the saturation level of total factor productivity;
\item pi, the labor participation rate (must be less than 1);
\item A0, the initial total factor productivity in year \code{start};
\item rho2, the carbon emissions intensity of technology 2;
\item rho3, the carbon emissions intensity of technology 3;
\item tau2, the half-saturation year of technology 2;
\item tau3, the half-saturation year of technology 3;
\item tau4, the half-saturation year of technology 4;
\item kappa, the rate of technological penetration;
\item aij, for $i,j=1, 2, 3$, the elements of the VAR cefficient matrix
(only when using a VAR model for the likelihood structure);
\item sigma_pop, the variance of the VAR or normal innovations for
population;
\item sigma_prod, the variance of the VAR or normal innovations for
economic output;
\item sigma_emis, the variance of the VAR or normal innovations for
emissions;
\item eps_pop, the variance of the observation errors for population
(only when using a VAR model for the likelihood structure);
\item eps_prod, the variance of the observation errors for economic
output (only when using a VAR model for the likelihood structure);
\item eps_emis, the variance of the observation errors for emissions.
(only when using a VAR model for the likelihood structure);
}}

\item{parnames}{Character vector of parameter names. These names should
align with the values in \code{pars}, but they don't need to be in any
particular order otherwise.}

\item{priors}{Named list of priors, with names containing all of the
elements in \code{parnames}. \code{priors} is best created using
'\code{\link{create_prior_list}}, but each element must be a list
containing: \itemize{
\item dens.fun, (string) function name for evaluating the density
\item ..., other named parameters required for evaluation of
\code{dens.fun}
}}

\item{dat}{List of data frames of data. List should have names 'pop'
(population), 'prod' (production), and 'emissions,' and each data frame
should have two columns, 'year' and 'value'.}

\item{lik_fun}{Character string with the name of the desired likelihood
function (provided functions are \code{log_lik_iid} and
'\code{log_lik_var}).}

\item{exp_gwp}{Boolean: should the provided expert assessment of average
GWP growth from 2010-2100 (\code{log_exp_gwp}) be inverted for additional
prior structure?}

\item{exp_co2}{Boolean: should the provided expert assessment of CO2
emissions in 2100 (\code{log_exp_co2}) be inverted for additional
prior structure?}

\item{exp_pop}{Boolean: should the provided expert assessment of population
in 2100 (\code{log_exp_pop}) be inverted for additional prior structure?}

\item{ff_thresh}{Numeric vector with two elements specifying the fossil fuel constraint thresholds. By default, this is NA so the constraint is turned off.}

\item{ff_const_yrs}{Numeric vector setting the years over which the fossil
fuel constraint should be evaluated. This can be a full sequence or a
vector with the start and end years.}

\item{ff_pen_windows}{List of matrices giving the bounds for a fossil fuel penetration constraint. Each list element should be a 2x3 matrix specifying the lower (row 1) and upper (row 2) penetration bounds for each year in which the constraint is specified. By default this is NA, so the constraint is turned off.}

\item{ff_pen_yrs}{Vector of years for which the penetration threshold is defined. By default this is NA.}

\item{hoyrs}{Integer vector of years that should be held out from the
likelihood function computation.}
}
\value{
Numeric value for the log-posterior of the parameters given the
 priors, the data and the fossil fuel constraint value.
}
\description{
\code{neg_log_post} returns the negative of the log-posterior density
 value for the provided parameter values, based on the provided list of
 prior distributions, the data, the fossil fuel constraint, and the
 desired likelihood structure. It is intended for use with an objective
 -minimizing optimization function such as \code{DEoptim}.
}
\details{
The provided prior distributions should be in the list form
 generated by \code{\link{create_prior_list}}.
}
