% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{residuals}
\alias{residuals}
\title{Calculate model residuals.}
\usage{
residuals(model_out, dat)
}
\arguments{
\item{model_out}{Data frame produced by model run, with columns 'year',
P' (population), 'Q' (production), and 'C' (emissions).}

\item{dat}{List of data frames of data. List should have names 'pop'
(population), 'prod' (production), and 'emissions,' and each data frame
should have two columns, 'year' and 'value'.}
}
\value{
List of model residuals, with the same names as the \code{dat}.
}
\description{
\code{residuals} computes the model residuals relative to the data.
}
