% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_model}
\alias{run_model}
\title{Run the integrated assessment model from Srikrishnan & Keller (2019).}
\usage{
run_model(pars, parnames, start = 1700, end = 2017, P0 = NULL,
  Q0 = NULL, C0 = NULL)
}
\arguments{
\item{pars}{Numeric vector of parameter values. These parameters must
include all of the model parameters listed in Table S1 of Srikrishnan &
Keller (2019), with Greek letters spelled out: \itemize{
\item psi1, the population growth rate;
\item psi2, the population half-saturation constant;
\item psi3, the population carrying capacity;
\item P0, the initial population in year \code{start};
\item lambda, the elasticity of production with respect to labor (must
be less than 1);
\item s, the savings rate;
\item delta, the capital depreciation rate (must be less than s);
\item alpha, the rate of technological progress for total factor
productivity;
\item As, the saturation level of total factor productivity;
\item pi, the labor participation rate (must be less than 1);
\item A0, the initial total factor productivity in year \code{start};
\item rho2, the carbon emissions intensity of technology 2;
\item rho3, the carbon emissions intensity of technology 3;
\item tau2, the half-saturation year of technology 2;
\item tau3, the half-saturation year of technology 3;
\item tau4, the half-saturation year of technology 4;
\item kappa, the rate of technological penetration.
}}

\item{parnames}{Character vector of parameter names. These names should
align with the values in \code{pars}, but they don't need to be in any
particular order otherwise.}

\item{start}{Numeric starting year for the simulation.}

\item{end}{Numeric ending year for the simulation.}

\item{P0}{Optional initial population value (if not treated as a
parameter).}

\item{Q0}{Optional initial economic output value. If not specified, the
initial value is treated initialized based on passed parameter values.}

\item{C0}{Optional initial carbon emissions value. If not specified,
the initial value is calculated based on the economic output and
emissions parameters.}
}
\value{
Data frame with columns \itemize{
 \item \code{year} (model years, between \code{start} and \code{end});
 \item \code{P} (population, in billions);
 \item \code{Q} (gross world product, or economic output, in trillions
 USD$2011);
 \item \code{C} (CO2 emissions, in Gt C/yr).}
}
\description{
\code{run_model} runs the integrated assessment model over the specified
   period (from start to end).
}
\details{
This function runs the model using the specified parameters. It can be
 started with arbitrary initial conditions, though these are detected
 based on whether or not \code{P0} (as a function argument) is
 NULL.
}
