% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sim_model}
\alias{sim_model}
\title{Simulate IAM based on provided parameters.}
\usage{
sim_model(pars, parnames, type, start = 1700, end = 2100, P0 = NULL,
  Q0 = NULL, C0 = NULL)
}
\arguments{
\item{pars}{Numeric vector of parameter values. These parameters must
include all of the model parameters listed in Table S1 of Srikrishnan &
Keller (2019), with Greek letters spelled out: \itemize{
\item psi1, the population growth rate;
\item psi2, the population half-saturation constant;
\item psi3, the population carrying capacity;
\item P0, the initial population in year \code{start};
\item lambda, the elasticity of production with respect to labor (must
be less than 1);
\item s, the savings rate;
\item delta, the capital depreciation rate (must be less than s);
\item alpha, the rate of technological progress for total factor
productivity;
\item As, the saturation level of total factor productivity;
\item pi, the labor participation rate (must be less than 1);
\item A0, the initial total factor productivity in year \code{start};
\item rho2, the carbon emissions intensity of technology 2;
\item rho3, the carbon emissions intensity of technology 3;
\item tau2, the half-saturation year of technology 2;
\item tau3, the half-saturation year of technology 3;
\item tau4, the half-saturation year of technology 4;
\item kappa, the rate of technological penetration;
\item aij, for $i,j=1, 2, 3$, the elements of the VAR cefficient matrix
(only when using a VAR model for the likelihood structure);
\item sigma_pop, the variance of the VAR or normal innovations for
population;
\item sigma_prod, the variance of the VAR or normal innovations for
economic output;
\item sigma_emis, the variance of the VAR or normal innovations for
emissions;
\item eps_pop, the variance of the observation errors for population
(only when using a VAR model for the likelihood structure);
\item eps_prod, the variance of the observation errors for economic
output (only when using a VAR model for the likelihood structure);
\item eps_emis, the variance of the observation errors for emissions.
(only when using a VAR model for the likelihood structure);
}}

\item{parnames}{Character vector of parameter names. These names should
align with the values in \code{pars}, but they don't need to be in any
particular order otherwise.}

\item{type}{String with the type of residual structure used for the
statistical noise. This can be "iid," "var," or "none" (to just return
model output with no noise).}

\item{start}{Starting year for the model (by default, 1700).}

\item{end}{Ending year for the model.}

\item{P0}{Optional initial value for global population (in billions).}

\item{Q0}{Optional initial value for gross world product (in trillions
USD2011)}

\item{C0}{Optional initial value for carbon emissions (in Gt C/yr).}
}
\value{
Data frame with columns "year" (model year), "P" (global
 population in billions), "Q" (gross world product in trillions USD2011),
 and "C" (carbon emissions in Gt C/yr).
}
\description{
\code{sim_model} simulates the IAM, including statistical noise, over the
 desired period of time. If initial parameters P0, Q0, and C0 are passed,
 the model will be initialized with them; if they are NULL, the model will
 be initialized as in Srikrishnan and Keller (2019), with an uncertain
 initial population.
}
