% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{var.sim}
\alias{var.sim}
\title{Simulate VAR(1) noise.}
\usage{
var.sim(N, A, W)
}
\arguments{
\item{N}{Length of VAR(1) noise sequence.}

\item{A}{VAR coefficient (square) matrix.}

\item{W}{VAR innovation (square) matrix.}
}
\value{
Matrix of generated noise samples. Each row corresponds to a
 different variable.
}
\description{
\code{var.sim} simulates vector autoregressive of order 1 (VAR(1)) noise.
}
